Piro: in transient solver, copy X into gx_out only if compatible
It is possible that X space has changed during integration,
if mesh adaptation was used. In this case, warn the user,
and have them taking care of this particular step.

Note that all dgd* entries still cannot be computed for
any response whose space is linked to the space of X. But
the sensitivities in TransientSolver are only computed for
responses 0,...,num_g-1, which do not include gx_out.
All other responses require dgdx and dxdp, which are taken
from the integrator (rather than from outArgs), so the
underlying app can make sure they have the right space.

Signed-off-by: Luca Bertagna <lbertag@sandia.gov>